Implement pruning of dead targets.
Now that ARPC provides a GetState() function, extend listeners to have
an IsDead() function. Extend Directory to have a PruneDeadTargets()
function that is called prior to iteration.

While there, remove locking of channels as done by the listeners. GRPC
channels are thread-safe, so we should do the same thing for ARPC.