Use a streaming RPC to implement List().
Even though we want to copy out the entire list of targets in class
Directory to simplify locking, it is advantageous to stream the results
back to the client. This allows us to stream back an arbitrary number of
targets. Only the labels of a target can now restrict us from obtaining
certain entries.